+++ src/tiled/automapper.h
@@ -35,6 +35,7 @@
 #include <QVector>
 
 #include <memory>
+#include <optional>
 #include <type_traits>
 #include <unordered_map>
 #include <vector>
@@ -260,7 +261,7 @@ public:
         /**
          * Determines whether the rules on the map need to be matched in order.
          */
-        bool matchInOrder = false;
+        std::optional<bool> matchInOrder;
 
         /**
          * This variable determines, how many overlapping tiles should be used.